refactor(indexer): split chunk-only indexing modules
